refactor(cli): extract prompt-for-config + reorder fns in init file
